Move Backend files into a backend folder
