--- AdjutantApi.Data/AdjutantContext.cs
@@ -1,13 +1,0 @@
-using AdjutantApi.Data.Models;
-using Microsoft.EntityFrameworkCore;
-
-namespace AdjutantApi.Data
-{
-    public class AdjutantContext : DbContext
-    {
-        public DbSet<VerificationKey> VerificationKeys { get; set; }
-        public DbSet<DiscordAccount> DiscordAccounts { get; set; }
-        
-        public AdjutantContext(DbContextOptions<AdjutantContext> options) : base(options){}
-    }
-}
