--- AdjutantApi.Data/Models/KeyConsumptionState.cs
@@ -1,7 +1,0 @@
-namespace AdjutantApi.Data.Models
-{
-    public enum KeyConsumptionState
-    {
-        Unconsumed, Consumed, Banned
-    }
-}
