--- AdjutantApi.Data/Models/VerificationKey.cs
@@ -1,13 +1,0 @@
-using System.ComponentModel.DataAnnotations;
-
-namespace AdjutantApi.Data.Models
-{
-    public class VerificationKey
-    {
-        public int Id { get; set; }
-        public string KeyValue { get; set; }
-        public KeyConsumptionState ConsumptionState { get; set; }
-        public string DisplayName { get; set; }
-    }
-}
-
