--- AdjutantApi/ViewModels/DisplayNameUpdate.cs
@@ -1,8 +1,0 @@
-namespace AdjutantApi.ViewModels
-{
-    public class DisplayNameUpdate
-    {
-        public string KeyValue { get; set; }
-        public string NewDisplayName { get; set; }
-    }
-}
