+++ backend/AdjutantApi.Data/Models/KeyConsumptionState.cs
@@ -1,0 +1,7 @@
+namespace AdjutantApi.Data.Models
+{
+    public enum KeyConsumptionState
+    {
+        Unconsumed, Consumed, Banned
+    }
+}
