+++ backend/AdjutantApi/ViewModels/DisplayNameUpdate.cs
@@ -1,0 +1,8 @@
+namespace AdjutantApi.ViewModels
+{
+    public class DisplayNameUpdate
+    {
+        public string KeyValue { get; set; }
+        public string NewDisplayName { get; set; }
+    }
+}
